# src/main/resources/api/alertagent-openapi.yml
swagger: '2.0'
info:
  title: alertagent OpenAPI
  version: 2.0.0
  description: alertagent OpenAPI
  license:
    name: 优云软件
    url: 'https://www.uyun.cn'

# 上下文路径（禁止通过 server.servlet.context-path 配置）
basePath: '/alertagent/openapi/v1'

# API类型（forntapi | serviceapi | openapi）
x-api-type: openapi

# 应用类型，与 spring.application.name 配置保持一致
x-application-name: alertagent-main

schemes:
  - http

produces:
  - application/json

tags:
  - name: rotatodo
    description: 值班待办

paths:
  '/todos/query':
    get:
      operationId: queryTodos
      summary: '查询待办事项'
      description: '查询待办事项'
      tags:
        - rotatodo
      parameters:
        - name: apikey
          in: query
          description: 访问APIKEY
          type: string
          required: true
        - name: dutyDate
          in: query
          description: '排班日期'
          required: true
          type: integer
          format: int64
        - name: currentUser
          in: query
          description: '当前用户'
          required: true
          type: string
      responses:
        '200':
          description: 成功
          schema:
            type: array
            items:
              $ref: '#/definitions/TodoDTO'
        400:
          description: 失败
          schema:
            $ref: '#/definitions/ErrorResponse'

definitions:
  ErrorResponse:
    type: object
    required:
      - error
    properties:
      error:
        $ref: '#/definitions/Error'
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: string
        description: 错误编码
        example: ERR-500
      message:
        type: string
        description: 错误消息
        example: 用户名称不能为空
      path:
        type: string
        description: 请求路径
        example: '/v2/cloud/save'
      time:
        type: string
        description: 出错时间
        example: '2019-03-23 09:11:20'
  TodoDTO:
    type: object
    properties:
      id:
        type: integer
        description: 'id'
      createUser:
        type: string
        description: '创建用户'
      dealsUser:
        type: string
        description: '处理用户'
      hasCompleted:
        type: boolean
        description: '是否已经完成'
      createTime:
        type: integer
        format: int64
        description: '创建时间'
      content:
        type: string
        description: '待办内容'
      todoTime:
        type: integer
        format: int64
        description: '待办时间'
      dealsTime:
        type: integer
        format: int64
        description: '处理时间'